Make the default policy 1w:5
There are >4 weeks in the majority of months. For that reason, some other tools
(e.g. [AutoMySQLBackup](http://www.debianhelp.co.uk/mysqlscript.htm),
[AutoPostgresqlBackup](https://launchpad.net/ubuntu/+source/autopostgresqlbackup/1.0-1))
seem to default to 5 weekly backups.

I think the worst possible scenario is as follows - Your monthly rotation
happens on the 1st of the month. It's now the 31st of the month. You have the
default strategy. On a calendar that looks something like this:

```
30      6       13      20      27 d
31      7       14      21      28 d
1  m    8       15      22      29 d
2       9       16      23      30 d
3       10 w    17 w    24 w    31 dw <- today
4       11      18      25 d
5       12      19      26 d
```

It's now the 31st of the month and you might expect that there is no more than
7 days between your files (as you have weekly rotation) for the past month.
However note that there is a 9 day period between monthly rotated file taken on
the 1st of the month and your last weekly rotation which happens to be on the
3,10,17,24,31 days of this month - but the rotation on the 3rd is being aged out.

For this reason, I suggest making the default 1w:5