Remove -format flag
The usefulness of the -format flag is questionable and complicates the
interface somewhat, especially given that it uses the Go convention of
expecting a date representing January 2, 2006.

Remove the -format flag and enforce ISO8601.

This also happily removes a call to `panic()` which should have been
removed earlier.

See discussion:
https://twitter.com/mattbostock/status/633030902327701505